refactor(docentes): changed the docentes modal to fit the new requirements
