+++ packages/shared/src/components/Buttons/ButtonSimple.jsx
@@ -4,17 +4,19 @@ import { Typography, Grid } from '@mui/material';
 import ButtonUnstyled from '@mui/base/ButtonUnstyled';
 import '../../styles/buttons/ButtonAdd.css';
 
-export default function ButtonSimple({ text, onClick, align }) {
+export default function ButtonSimple({
+  text, onClick, align, design,
+}) {
   const justifyContent = {
     left: 'flex-start',
     center: 'center',
     right: 'flex-end',
-  }[align];
+  }[align] || 'flex-start';
 
   return (
     <Grid container justifyContent={justifyContent}>
       <Grid item>
-        <ButtonUnstyled className="buttonAdd guardar" onClick={onClick}>
+        <ButtonUnstyled className={`buttonAdd ${design}`} onClick={onClick}>
           <Typography variant="body1">{text}</Typography>
         </ButtonUnstyled>
       </Grid>
@@ -24,10 +26,13 @@ export default function ButtonSimple({ text, onClick, align }) {
 
 ButtonSimple.defaultProps = {
   align: 'left',
+  design: 'guardar',
+  onClick: () => {},
 };
 
 ButtonSimple.propTypes = {
-  onClick: PropTypes.func.isRequired,
+  onClick: PropTypes.func,
   text: PropTypes.string.isRequired,
+  design: PropTypes.string,
   align: PropTypes.oneOf(['left', 'center', 'right']),
 };
