+++ packages/solicitudes/src/index.js
@@ -18,7 +18,7 @@ import modalidades from './components/utils/Mocks/mockModalidades';
 import formData from './components/utils/sections/forms/formData';
 import getSolicitudDetalles from './components/utils/getSolicitudDetalles';
 import columnsSolicitudes from './components/utils/Tables/solicitudesColumns';
-import DocentesCreateModal from './components/utils/Components/DocentesModales/DocentesCreateModal';
+import DocentesCreateModal from './components/utils/Components/DocentesModales/DocentesModal';
 
 export {
   getSolicitudDetalles,
